requirements-heroku.txt
django-toolbelt==0.0.1
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
